Deactivate output if thermocouple in error state
This is a safety measure to prevent the output from ever being stuck on
in the case that the thermocouple is unplugged / malfunctions. If the
thermocouple reports an error state, we set output power to zero.

This should also make our control more resilient to the intermittent
thermocouple 'open-circuit' error states I've been seeing. The error
state doesn't stop the temperature callback from coming through: it's
called over and over with the same value. This causes the control to
overcompensate. By keeping track of the error states, we can avoid
calling the PID with these erroneous values