Add recovery time from error state
At very high temperatures (>950C), the thermocouple is reported as going
open circuit intermittently. This seems to be caused by some kind of
interferrence, and usually recovers within 5 seconds or so. Disabling
the output as soon as we hit this state results in poor performance at
these high temperatures since the whole PID loop is interrupted.

So here we only deactivate the output if the thermocouple is reporting
an error state for an extended period (over 11 seconds). This will still
provide safety in the case of real thermocouple failure but will be less
of a hair-trigger at the higher temperatures.